# backend-master/docs/apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BISMA API",
    "description": "API Documentation For Bisma",
    "version": "1.0.0"
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/positions": {
      "get": {
        "tags": ["Positions"],
        "summary": "Get list of positions",
        "description": "Retrieve a list of positions with optional filtering and pagination",
        "operationId": "listPosisi",
        "parameters": [
          {
            "name": "posisi",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Manager"
            },
            "description": "Filter positions by name (optional)"
          },
          {
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Number of items per page"
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Current page number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posisiList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "posisi": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "total_item": {
                      "type": "integer"
                    },
                    "total_page": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No positions found"
          },
          "400": {
            "description": "Invalid request parameters"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Positions"],
        "summary": "Create Positions",
        "consumes": "application/json",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "posisi": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/department": {
      "get": {
        "tags": ["Department"],
        "summary": "Get list of departemen",
        "description": "Retrieve a list of departemen with optional filtering and pagination",
        "operationId": "listPosisi",
        "parameters": [
          {
            "name": "department",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Ilmu Komputer"
            },
            "description": "Filter departemen by name (optional)"
          },
          {
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Number of items per page"
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Current page number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posisiList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "posisi": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "total_item": {
                      "type": "integer"
                    },
                    "total_page": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No positions found"
          },
          "400": {
            "description": "Invalid request parameters"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Department"],
        "summary": "Create Department",
        "consumes": "application/json",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "department": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}